SDK-1659: Split out yoti_client_test into respective aml, profile & attribute tests, use "assert.NilError" instead of assert "is.Nil"
