+++ test/decode_test_file.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"gotest.tools/v3/assert"
-	is "gotest.tools/v3/assert/cmp"
 )
 
 // DecodeTestFile reads a test fixture file
@@ -14,7 +13,7 @@ func DecodeTestFile(t *testing.T, filename string) (result []byte) {
 	base64String := string(base64Bytes)
 	filebytes, err := base64.StdEncoding.DecodeString(base64String)
 
-	assert.Assert(t, is.Nil(err))
+	assert.NilError(t, err)
 
 	return filebytes
 }
